requirements.txt
python-decouple
validators
svgwrite
psycopg2
gunicorn